Model supports multiple errors, muliple view methods
Errors
------

Eliminates use of `log.Fatalln` inside terminal application.

Any potential error will be appended to a slice of errors. Any non-nil
errors with be displayed in the application.

Errors that the application can perceivably tolerate are ignored. The
impact would be missing data or null values. Such defects are not
considered particularly problematic. It's a bit lazy, but adequate for
now.

Veiws
-----

`errorView`, `confirmDeleteView`, `thingView`, `typeView` created to
simplify the `View` method.